split index-item into own view. add own version detection.
